Extend from a base view.
All views will need to marshal the data before it does anything with it so lets do that in a base view.
