+++ lib/views/Delta.js
@@ -1,21 +1,24 @@
 var formatter = require('../utils/formatter'),
   moment = require('moment-timezone'),
-  _ = require('lodash');
+  _ = require('lodash'),
+  View = require('./View'),
+  util = require('util');
 
 require('../utils/date-functions');
 
-function Delta (module) {
-  this.moduleConfig = module.moduleConfig;
-  this.data = module.dataSource.data;
-  this.axes = module.axes;
-
-  this.createDeltas();
+function Delta () {
+  this.initialize.apply(this, arguments);
 }
+
+util.inherits(Delta, View);
+
+Delta.prototype.initialize = function () {
+  View.prototype.initialize.apply(this, arguments);
+  this.createDeltas();
+};
 
 Delta.prototype.createDeltas = function () {
   var yAxes = this.axes.y,
-      reverseData = false,
-      groupedData = false,
       calculateDelta = canCalculateDelta(yAxes);
 
   this.formating = this.moduleConfig.format || this.moduleConfig['format-options'];
@@ -27,70 +30,19 @@ Delta.prototype.createDeltas = function () {
     calculateDelta = true;
   }
 
-  if (this.moduleConfig['module-type'] !== 'grouped_timeseries') {
-    //For a delta we're only interested in the first axis for everything but grouped_timeseries
-    yAxes = [yAxes[0]];
-  }
-
-  if (this.moduleConfig['data-source'] && this.moduleConfig['data-source']['query-params']) {
-
-    if (this.moduleConfig['data-source']['query-params'].period) {
-      reverseData = true;
-    }
-
-    if (this.moduleConfig['data-source']['query-params'].group_by && yAxes.length > 1) {
-      _.each(this.moduleConfig['data-source']['query-params'].group_by, function (group, index) {
-        if (group) {
-          if (index === 0) {
-            this.data = _.groupBy(this.data, group);
-          } else {
-            var newData = {};
-            _.each(this.data, function (dataGroup, groupKey) {
-              var tmpGroup = _.groupBy(dataGroup, group);
-              _.each(tmpGroup, function (dataSubgroup, subgroupKey) {
-                newData[groupKey + ':' + subgroupKey] = dataSubgroup;
-              });
-            });
-            this.data = newData;
-          }
-          groupedData = true;
-        }
-      }, this);
-    }
-
-    if (this.moduleConfig['show-total-lines']) {
-      var valueAttr = this.moduleConfig['value-attribute'];
-      //create a total grouping
-      var groups = Object.keys(this.data);
-      var totals = [];
-      _.each(groups, function (group, index) {
-        _.each(this.data[group], function (item, groupIndex) {
-          if (index === 0) {
-            totals.push(_.clone(item));
-          } else {
-            if (item[valueAttr]) {
-              totals[groupIndex][valueAttr] += item[valueAttr];
-            }
-          }
-        }, this);
-      }, this);
-      this.data.total = totals;
-    }
-  }
-
   _.each(yAxes, function (yAxis) {
     var valueAttr = yAxis.key || this.moduleConfig['value-attribute'];
     var groupId = yAxis.groupId;
     var data;
 
-    if (groupedData) {
+    if (this.groupedData) {
       data = this.data[groupId];
       calculateDelta = true;
     } else {
       data = this.data;
     }
 
-    if (reverseData) {
+    if (this.reverseData) {
       data.reverse();
     }
