+++ lib/views/Table.js
@@ -1,17 +1,22 @@
 var formatter = require('../utils/formatter'),
-  _ = require('lodash');
+  _ = require('lodash'),
+  View = require('./View'),
+  util = require('util');
 
-function Table (module, options) {
+function Table () {
+  this.initialize.apply(this, arguments);
+}
+
+util.inherits(Table, View);
+
+Table.prototype.initialize = function (module, options) {
+  View.prototype.initialize.apply(this, arguments);
   this.options = _.extend({
     colsLimit: false,
     rowsLimit: false
   }, options);
-  this.moduleConfig = module.moduleConfig;
-  this.data = module.dataSource.data;
-  this.axes = module.axes;
-
   this.tabularise();
-}
+};
 
 Table.prototype.render = function () {
   var output = [];
@@ -41,25 +46,6 @@ Table.prototype.tabularise = function () {
   var needsTotals = false;
   var yAxes = this.axes.y;
   var xCol = [this.axes.x.label];
-
-  if (this.moduleConfig['data-source']['query-params'].group_by && yAxes.length > 1) {
-    _.each(this.moduleConfig['data-source']['query-params'].group_by, function (group, index) {
-      if (group) {
-        if (index === 0) {
-          this.data = _.groupBy(this.data, group);
-        } else {
-          var newData = {};
-          _.each(this.data, function (dataGroup, groupKey) {
-            var tmpGroup = _.groupBy(dataGroup, group);
-            _.each(tmpGroup, function (dataSubgroup, subgroupKey) {
-              newData[groupKey + ':' + subgroupKey] = dataSubgroup;
-            });
-          });
-          this.data = newData;
-        }
-      }
-    }, this);
-  }
 
   _.each(yAxes, function (yAxis, index) {
     var yCol = [];
